First attempt at implementing data structures
Having got the hang of pointers, I'm really enjoying programming in C, however often I find myself drifting to python for ease of use. In order to avoid this, I've decided to try and implement my own library of C commands, to increase the complexity of the projects I can produce in C. More soon.