A dynamic problem solved using divide and conquer.
This was some code that I wrote a few weeks back, before I had got the hang of coding in C, and an understanding of algorithm paradigms. The question was, given an array of digits, how many ways could they be summed so that a certain result could be obtained. I, naively implemented a recursive program to repeatedly call the function on the input minus one value in the array. It works, but looking back, I can't help but mentally flinch whenever I look at this code.