Regex to NFA Parser
Wanted to build a transpiler - turns out the first step is building a lexical analyser. And the best way to do that is to build a finite state automata - to do that will need a parser to turn regex into discreet state automata. So I've implemented the first step - something that takes a regular expression and produces an NFA - next step is to convert the NFA into a DFA. And then we'll be cooking with fire.